Force SSL in production
