+++ packages/backend/src/functions/workspace/initializeRandomDirectory.ts
@@ -1,0 +1,9 @@
+import { promises as fs } from "fs";
+import { sep } from "path";
+import os from "os";
+
+export async function initializeRandomDirectory(): Promise<string> {
+  const tmpDir = os.tmpdir();
+  const projectRoot = await fs.mkdtemp(`${tmpDir}${sep}`);
+  return projectRoot;
+}
